backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
polyline==2.0.0